Speed up compilation a ton.
tsc hadn't been told to ignore its output directory so it would try to
build them as well and give errors. To get around this we always
cleaned the output directory.

I've instead made it so that tsc knows not to build its output
directory. This also lets us use `--build` instead of `--project` which
checks the modified dates of its input and output to see if anything
needs to be done.